refactor(#3479): extract tmux_watcher utf8 decoder + terminal-readiness helpers to submodules
Phase-1 rank-2 PURE MOVE (zero logic change) shrinking the frozen giant
src/services/discord/tmux_watcher.rs from 7485 -> 7241 production LoC by
extracting two cohesive pure clusters into tmux_watcher/ child modules:

- utf8_chunk_decoder.rs (88 prod LoC): Utf8ChunkDecoder + DecodedUtf8Chunk
  streaming UTF-8 decoder that buffers a split trailing multibyte scalar
  across read boundaries. Self-contained (std-only).
- terminal_readiness.rs (214 prod LoC): the synchronous terminal-readiness /
  inflight-classification predicates (adopt_watcher_terminal_message_ids_from_inflight,
  watcher_inflight_represents_external_input/_is_panel_eligible/
  _needs_anchor_lifecycle_cleanup, watcher_direct_terminal_should_commit_session_idle,
  watcher_terminal_token_update_status, the JSONL ready_for_input sentinel probes)
  plus the pure discard_watcher_pending_buffer_after_suppressed_turn reconciler.

The async, shared-touching commit_watcher_direct_terminal_session_idle (which
sits BETWEEN the two readiness clusters in the root) deliberately STAYS in
tmux_watcher.rs — it couples to shared.mailbox/core + internal_api .await.

Items are pub(super) and re-imported via `use self::X::*` so the watcher loop's
call sites stay byte-identical; the sibling panel_decisions module keeps
resolving watcher_inflight_is_panel_eligible via the parent re-export glob.
Moved unit tests live in sibling *_tests.rs (excluded from the production
splitter). Lowered the giant ratchet baseline 7485 -> 7241 and registered the
two new child modules; refreshed change-surfaces.md + generated inventory docs.
